Create a base trait for the resource controllers.
